Simplify GPU Memory Service module boundaries
