+++ lib/gpu_memory_service/common/protocol/messages.py
@@ -3,25 +3,10 @@
 
 """Message types for GPU Memory Service RPC protocol."""
 
-from enum import Enum
 from typing import List, Optional, Union
 
 import msgspec
-
-
-class RequestedLockType(str, Enum):
-    """Lock type requested by client."""
-
-    RW = "rw"
-    RO = "ro"
-    RW_OR_RO = "rw_or_ro"
-
-
-class GrantedLockType(str, Enum):
-    """Lock type actually granted by server."""
-
-    RW = "rw"
-    RO = "ro"
+from gpu_memory_service.common.locks import GrantedLockType, RequestedLockType
 
 
 class HandshakeRequest(msgspec.Struct, tag="handshake_request"):
