+++ lib/gpu_memory_service/integrations/vllm/model_loader.py
@@ -14,10 +14,12 @@ import logging
 from typing import TYPE_CHECKING
 
 import torch
-from gpu_memory_service import get_or_create_gms_client_memory_manager
-from gpu_memory_service.client.torch.allocator import gms_use_mem_pool
+from gpu_memory_service.client.torch.allocator import (
+    get_or_create_gms_client_memory_manager,
+    gms_use_mem_pool,
+)
 from gpu_memory_service.client.torch.module import materialize_module_from_gms
-from gpu_memory_service.common.types import GrantedLockType
+from gpu_memory_service.common.locks import GrantedLockType
 from gpu_memory_service.common.utils import get_socket_path
 from gpu_memory_service.integrations.common.utils import (
     finalize_gms_write,
