+++ lib/gpu_memory_service/server/session.py
@@ -1,204 +1,62 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Server-side connection, FSM, and waiter state."""
+"""Server-side lock acquisition and cleanup."""
 
 from __future__ import annotations
 
 import asyncio
-from dataclasses import dataclass, field
-from typing import Optional, Set
+from dataclasses import dataclass
+from typing import Optional
 
-from gpu_memory_service.common.types import (
-    RO_ALLOWED,
-    RW_ALLOWED,
-    RW_REQUIRED,
-    GrantedLockType,
-    RequestedLockType,
-    ServerState,
-    StateEvent,
+from gpu_memory_service.common.locks import GrantedLockType, RequestedLockType
+from gpu_memory_service.common.protocol.messages import (
+    AllocateRequest,
+    CommitRequest,
+    ExportAllocationRequest,
+    FreeAllocationRequest,
+    GetAllocationRequest,
+    GetAllocationStateRequest,
+    GetLockStateRequest,
+    GetStateHashRequest,
+    ListAllocationsRequest,
+    MetadataDeleteRequest,
+    MetadataGetRequest,
+    MetadataListRequest,
+    MetadataPutRequest,
 )
 
-
-@dataclass(eq=False)
-class Connection:
-    reader: asyncio.StreamReader
-    writer: asyncio.StreamWriter
-    mode: GrantedLockType
-    session_id: str
-    recv_buffer: bytearray = field(default_factory=bytearray)
-
-    def __hash__(self) -> int:
-        return hash(self.session_id)
-
-    async def close(self) -> None:
-        self.writer.close()
-        try:
-            await self.writer.wait_closed()
-        except Exception:
-            pass
-
-
-class InvalidTransition(Exception):
-    """Raised when an invalid state transition is attempted."""
+from .fsm import GMSFSM, Connection, ServerState, StateEvent
 
 
 class OperationNotAllowed(Exception):
-    """Raised when an operation is not allowed in the current state/mode."""
+    pass
 
 
-@dataclass(frozen=True)
-class Transition:
-    from_states: frozenset[ServerState]
-    event: StateEvent
-    to_state: Optional[ServerState]
-    condition: Optional[str] = None
+RW_REQUIRED: frozenset[type] = frozenset(
+    {
+        AllocateRequest,
+        FreeAllocationRequest,
+        MetadataPutRequest,
+        MetadataDeleteRequest,
+        CommitRequest,
+    }
+)
 
+RO_ALLOWED: frozenset[type] = frozenset(
+    {
+        ExportAllocationRequest,
+        GetAllocationRequest,
+        ListAllocationsRequest,
+        MetadataGetRequest,
+        MetadataListRequest,
+        GetLockStateRequest,
+        GetAllocationStateRequest,
+        GetStateHashRequest,
+    }
+)
 
-TRANSITIONS: list[Transition] = [
-    Transition(
-        from_states=frozenset({ServerState.EMPTY, ServerState.COMMITTED}),
-        event=StateEvent.RW_CONNECT,
-        to_state=ServerState.RW,
-    ),
-    Transition(
-        from_states=frozenset({ServerState.RW}),
-        event=StateEvent.RW_COMMIT,
-        to_state=ServerState.COMMITTED,
-    ),
-    Transition(
-        from_states=frozenset({ServerState.RW}),
-        event=StateEvent.RW_ABORT,
-        to_state=ServerState.EMPTY,
-    ),
-    Transition(
-        from_states=frozenset({ServerState.COMMITTED, ServerState.RO}),
-        event=StateEvent.RO_CONNECT,
-        to_state=ServerState.RO,
-    ),
-    Transition(
-        from_states=frozenset({ServerState.RO}),
-        event=StateEvent.RO_DISCONNECT,
-        to_state=ServerState.RO,
-        condition="has_remaining_readers",
-    ),
-    Transition(
-        from_states=frozenset({ServerState.RO}),
-        event=StateEvent.RO_DISCONNECT,
-        to_state=ServerState.COMMITTED,
-        condition="is_last_reader",
-    ),
-]
-
-
-class GMSLocalFSM:
-    """Explicit connection/lock state machine."""
-
-    def __init__(self):
-        self._rw_conn: Optional[Connection] = None
-        self._ro_conns: Set[Connection] = set()
-        self._committed = False
-
-    @property
-    def state(self) -> ServerState:
-        if self._rw_conn is not None:
-            return ServerState.RW
-        if self._ro_conns:
-            return ServerState.RO
-        if self._committed:
-            return ServerState.COMMITTED
-        return ServerState.EMPTY
-
-    @property
-    def rw_conn(self) -> Optional[Connection]:
-        return self._rw_conn
-
-    @property
-    def ro_conns(self) -> Set[Connection]:
-        return self._ro_conns
-
-    @property
-    def ro_count(self) -> int:
-        return len(self._ro_conns)
-
-    @property
-    def committed(self) -> bool:
-        return self._committed
-
-    def _has_remaining_readers(self, conn: Connection) -> bool:
-        return len(self._ro_conns) > 1 or conn not in self._ro_conns
-
-    def _is_last_reader(self, conn: Connection) -> bool:
-        return len(self._ro_conns) == 1 and conn in self._ro_conns
-
-    def _check_condition(self, condition: Optional[str], conn: Connection) -> bool:
-        if condition is None:
-            return True
-        if condition == "has_remaining_readers":
-            return self._has_remaining_readers(conn)
-        if condition == "is_last_reader":
-            return self._is_last_reader(conn)
-        raise ValueError(f"Unknown condition: {condition}")
-
-    def _find_transition(
-        self,
-        from_state: ServerState,
-        event: StateEvent,
-        conn: Connection,
-    ) -> Optional[Transition]:
-        for transition in TRANSITIONS:
-            if from_state not in transition.from_states:
-                continue
-            if transition.event != event:
-                continue
-            if not self._check_condition(transition.condition, conn):
-                continue
-            return transition
-        return None
-
-    def _apply_event(self, event: StateEvent, conn: Connection) -> None:
-        if event == StateEvent.RW_CONNECT:
-            self._rw_conn = conn
-            self._committed = False
-        elif event == StateEvent.RW_COMMIT:
-            self._committed = True
-            self._rw_conn = None
-        elif event == StateEvent.RW_ABORT:
-            self._rw_conn = None
-        elif event == StateEvent.RO_CONNECT:
-            self._ro_conns.add(conn)
-        elif event == StateEvent.RO_DISCONNECT:
-            self._ro_conns.discard(conn)
-
-    def transition(self, event: StateEvent, conn: Connection) -> ServerState:
-        transition = self._find_transition(self.state, event, conn)
-        if transition is None:
-            raise InvalidTransition(
-                f"No transition for {event.name} from state {self.state.name} "
-                f"(session={conn.session_id})"
-            )
-        self._apply_event(event, conn)
-        return self.state
-
-    def check_operation(self, msg_type: type, conn: Connection) -> None:
-        if conn.mode == GrantedLockType.RW and msg_type not in RW_ALLOWED:
-            raise OperationNotAllowed(
-                f"{msg_type.__name__} not allowed for RW session in state {self.state.name}"
-            )
-        if conn.mode == GrantedLockType.RO and msg_type not in RO_ALLOWED:
-            raise OperationNotAllowed(
-                f"{msg_type.__name__} not allowed for RO session in state {self.state.name}"
-            )
-        if msg_type in RW_REQUIRED and conn.mode != GrantedLockType.RW:
-            raise OperationNotAllowed(
-                f"{msg_type.__name__} requires RW session, got {conn.mode.value}"
-            )
-
-    def can_acquire_rw(self) -> bool:
-        return self._rw_conn is None and not self._ro_conns
-
-    def can_acquire_ro(self, waiting_writers: int) -> bool:
-        return self._committed and self._rw_conn is None and waiting_writers == 0
+RW_ALLOWED: frozenset[type] = RW_REQUIRED | RO_ALLOWED
 
 
 @dataclass(frozen=True)
@@ -215,7 +73,7 @@ class GMSSessionManager:
     """Owns lock transitions, waiter coordination, and cleanup."""
 
     def __init__(self):
-        self._locking = GMSLocalFSM()
+        self._locking = GMSFSM()
         self._waiting_writers = 0
         self._reserved_rw_session_id: Optional[str] = None
         self._condition = asyncio.Condition()
@@ -336,7 +194,18 @@ class GMSSessionManager:
         self._locking.transition(StateEvent.RW_COMMIT, conn)
 
     def check_operation(self, msg_type: type, conn: Connection) -> None:
-        self._locking.check_operation(msg_type, conn)
+        if conn.mode == GrantedLockType.RW and msg_type not in RW_ALLOWED:
+            raise OperationNotAllowed(
+                f"{msg_type.__name__} not allowed for RW session in state {self.state.name}"
+            )
+        if conn.mode == GrantedLockType.RO and msg_type not in RO_ALLOWED:
+            raise OperationNotAllowed(
+                f"{msg_type.__name__} not allowed for RO session in state {self.state.name}"
+            )
+        if msg_type in RW_REQUIRED and conn.mode != GrantedLockType.RW:
+            raise OperationNotAllowed(
+                f"{msg_type.__name__} requires RW session, got {conn.mode.value}"
+            )
 
     def begin_cleanup(self, conn: Optional[Connection]) -> StateEvent | None:
         if conn is None:
