+++ tests/gpu_memory_service/conftest.py
@@ -9,14 +9,14 @@ import importlib.util
 import sys
 from pathlib import Path
 
-# Prefer the in-repo GMS source tree when it exists, but let runtime images fall
-# back to the installed package when they do not ship /workspace/lib.
-for parent in Path(__file__).resolve().parents:
-    if (parent / "lib" / "gpu_memory_service").is_dir():
-        lib_dir = parent / "lib"
-        if str(lib_dir) not in sys.path:
-            sys.path.insert(0, str(lib_dir))
-        break
-else:
-    if importlib.util.find_spec("gpu_memory_service") is None:
+# Respect an already-resolvable GMS package first. This lets containerized runs
+# point pytest at an explicit source overlay without this hook overriding it.
+if importlib.util.find_spec("gpu_memory_service") is None:
+    for parent in Path(__file__).resolve().parents:
+        if (parent / "lib" / "gpu_memory_service").is_dir():
+            lib_dir = parent / "lib"
+            if str(lib_dir) not in sys.path:
+                sys.path.insert(0, str(lib_dir))
+            break
+    else:
         collect_ignore_glob = ["*"]
